fix(kvark): norske gruppetyper på medlemskap-kort, skjul TIHLDE
Medlemskap-kortene viste den rå UPPERCASE-typen fra databasen
(SUBGROUP, BOARD, …). Løft GROUP_TYPE_LABELS/groupTypeLabel ut av
grupper.tsx til delte lib/group.ts og bruk norsk visningsnavn på
kortene. Filtrer også bort TIHLDE-gruppen – det er ingen gruppe man
melder seg inn i, alle er med.
